% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Uni_EDA.R
\name{Categorical_Uni_EDA}
\alias{Categorical_Uni_EDA}
\title{Univariate Categorical Exploratory Data Analysis}
\usage{
Categorical_Uni_EDA(
  dataset,
  detect = T,
  ordinal.force = list(),
  nominal.force = c(),
  ignore.columns = c(),
  alternate.nas = list()
)
}
\arguments{
\item{dataset}{dataframe ; containing columns of each field and rows containing each record}

\item{detect}{Boolean ; True or False value on where the user wishes the function to automatically detect additional categorical fields, if false will only use the forced columns}

\item{ordinal.force}{list ; containing vectors for each column you wished forced and analysed as ordinal data type. The first element of each vector should be the name/index of the column you wish to force, followed by levels of the ordinal data in order you wish them to be handled.}

\item{nominal.force}{vector ; containing the names of the columns (as strings) or indexes of columns, or a combination, that should be forced and analysed as nominal data type}

\item{ignore.columns}{vector ; the names of the columns (as strings) or indexes of columns, or a combination, that should be ignored during analysis}

\item{alternate.nas}{list ; containing vectors for each column you wish to specify alternate/additional NA values for. The first element of each vector should be the name/index of the column you wish to force followed by the additional values in the column that should be considered as NA.}
}
\value{
A list containing:
\itemize{
\item Nominal_Data : Adjusted nominal data that was analysed
\item Ordinal_Data: Adjusted ordinal data that was analysed
\item Ordinal_Statistics: List containing the most and least common categories along with their counts for ordinal data
\item Nominal_Statistics: List containing the most and least common categories along with their counts for nominal data
\item Nominal_Counts: List containing the counts for each category in the nominal data
\item Ordinal_Counts: List containing the counts for each category in the ordinal data
}
}
\description{
This function receives a dataframe containing survey data with one column
for each question and will output summary and characteristic values for the
categorical data. Analysis is split in to ordinal and nominal types. Columns
can either be automatically detected or forced by the user for analysis.
}
\examples{
Categorical_Uni_EDA(dataset = example_dataset,
                    detect = T,
                    ordinal_force = list(c('col1', 'a','b','c'), c('col2', 2,3,4)),
                    nominal_force = c('col5','col7'),
                    ignore.columns = c('col4'),
                    alternate.nas = list(c('col1', 'error'))

Categorical_Uni_EDA(example_dataset)

plot(Categorical_Uni_EDA(example_dataset))

print(Categorical_Uni_EDA(example_dataset))

}
