% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Describe.R
\name{Data_Describe}
\alias{Data_Describe}
\title{Data Describe}
\usage{
Data_Describe(
  dataset,
  NLP.force = c(),
  ordinal.force = list(),
  nominal.force = c(),
  numeric.force = c(),
  alternate.nas = list(),
  preserve.nonconform = T,
  remove.repeat = T
)
}
\arguments{
\item{dataset}{dataframe ; containing columns of each field and rows containing each record}

\item{NLP.force}{vector ; containing the names of the columns (as strings) or indexes of columns, or a combination, that should be forced and analysed as natural language (NLP) data type}

\item{ordinal.force}{list ; containing vectors for each column you wished forced and analysed as ordinal data type. The first element of each vector should be the name/index of the column you wish to force, followed by levels of the ordinal data in order you wish them to be handled.}

\item{nominal.force}{vector ; containing the names of the columns (as strings) or indexes of columns, or a combination, that should be forced and analysed as nominal data type}

\item{numeric.force}{vector ; containing the names of the columns (as strings) or indexes of columns, or a combination, that should be forced and analysed as numeric data type}

\item{alternate.nas}{list ; containing vectors for each column you wish to specify alternate/additional NA values for. The first element of each vector should be the name/index of the column you wish to force followed by the additional values in the column that should be considered as NA.}

\item{preserve.nonconform}{boolean; Default T - true of false as to whether to analyse potentially natuaral language answers within other data type columns and to split these answers in to a separate column}

\item{remove.repeat}{boolean;  Default T - true or false as to whether to remove repeat rows within the dataset}
}
\value{
A list containing summary statistics about the dataset as a whole. These include
\itemize{
\item dimensions: dataframe; dimensions for the original input dataset as well as the adjusted output dataset
\item memory_size: numeric; the memory size in bites that the dataset takes up
\item data_type_conversion: dataframe; Contains each datafield, the original input data type as well as the detected/output datatype
\item data_type_names: dataframe; a dataframe containing a column per datatype and the columns containing names of the fields of that datatype
\item missingness_percentage: vector; A vector containing the percetage of missing data foudn in each data field
\item grouped_missingness: dataframe; Dataframe containing columns of the names of each data field with columns split in to complete, 0-10\%, 10-25\%, 25-50\%, 50-75\% and >75\% missing data
\item repeat_rows: numeric; the number of repeat rows found in the dataset
\item data: dataframe; the output adjusted data.
}
}
\description{
Data Describe
}
\details{
For details on auto detection of data type see documentation for data_type_detect() function.
}
\examples{
Data_Describe(dataset = example_dataset,
                    NLP.force = list(c('col1', 'a','b','c'), c('col2', 2,3,4)),
                    ordinal_force = list(c('col3', 'a','b','c'), c('col4', 2,3,4)),
                    nominal_force = c('col5','col6'),
                    numeric.force = c('col7','col8'),
                    preserve.nonconform = T,
                    alternate.nas = list(c('col1', 'error')),
                    remove.repeat = T)

Data_Describe(example_dataset)


plot(Data_Describe(example_dataset))

print(Data_Describe(example_dataset))

}
