% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natural_language_column}
\alias{natural_language_column}
\title{University of Waterloo Course Reviews (natural_language_column)}
\format{
\subsection{\code{natural_language_column}}{

A data frame with 200 rows and 12 columns:
\describe{
\item{reviews}{Top 1000 reviews from the Waterloo course reviews dataset}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/anthonysusevski/course-reviews-university-of-waterloo}
}
\usage{
natural_language_column
}
\description{
Top 1000 reviews from University of Waterloo course reviews dataset on kaggle
}
\keyword{datasets}
