% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Column_Analysis.R
\name{NLP_Column_Analysis}
\alias{NLP_Column_Analysis}
\title{Analysis of a Single Natural Language Vector}
\usage{
NLP_Column_Analysis(column_in, input_name)
}
\arguments{
\item{column_in}{vector; A vector of character entries, each entry should be of natural language format data}

\item{input_name}{character; The name of the field being analysed}
}
\value{
A list containing the following analysis:
\itemize{
\item Word_Frequency: dataframe; containing the top 500 most common words and their count
\item Bigrams: dataframe; containing the top 500 most common bigrams (word pairings) from the data and their frequency (order specific)
\item Trigrams: dataframe; containing the top 500 most commonly mentioned trigrams (three words mentioned consecutively) and their counts (order specific)
\item Corr_matrix: dataframe; containing the pearson frequency corrolation for words mentioned in the same response (order non specific)
\item Tf_idf: tibble; containing the word, which response and tf_idf score for the analysis. This analysis works better on longer response answers
\item wordcount: dataframe; contains the word count for each response
\item Sentiments: dataframe; dataframe containing scores for anger, anticipation, disgust, fear, joy, sadness, surprise, trust, negative and positive for each response. This is determined using the NRC sentiment dictionary.
\item intput_name: character; The name of the data field that the analysis refers to.
}
}
\description{
Analysis of a Single Natural Language Vector
}
\examples{

NLP_Column_Analysis('Column1', 'Question_2')

plot(NLP_Column_Analysis(c(1), 'Question_2'))
print(NLP_Column_Analysis(c(1), 'Question_2'))

}
