% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Type_Detect.R
\name{Numeric_Type_Detect}
\alias{Numeric_Type_Detect}
\title{Title}
\usage{
Numeric_Type_Detect(input_vector, dataset, preserve_nonconform = T, force = F)
}
\arguments{
\item{force}{}
}
\description{
Title
}
