% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Uni_EDA.R
\name{Numeric_Uni_EDA}
\alias{Numeric_Uni_EDA}
\title{Univariate Numeric Exploratory Data Analysis}
\usage{
Numeric_Uni_EDA(
  dataset,
  numeric.force = c(),
  ignore.columns = c(),
  alternate.nas = list()
)
}
\arguments{
\item{dataset}{dataset dataframe ; containing columns of each field and rows containing each record}

\item{numeric.force}{vector ; containing the names of the columns (as strings) or indexes of columns, or a combination, that should be forced and analysed as numeric data type}

\item{ignore.columns}{vector ; the names of the columns (as strings) or indexes of columns, or a combination, that should be ignored during analysis}

\item{alternate.nas}{list ; containing vectors for each column you wish to specify alternate/additional NA values for. The first element of each vector should be the name/index of the column you wish to force followed by the additional values in the column that should be considered as NA.}
}
\value{
List containing:
\enumerate{
\item Integer_characteristics - A list containing:
}
\itemize{
\item Means: vector; Mean of each integer field
\item Medians: vector; Medians of each integer field
\item Modes: list; list of a vector per field containing the modal integer values for each field
\item Maxs: vector; max of each integer field
\item Mins: vector; Minimum of each integer field
\item I.Q.Rs: vector; Inter quartile range of each integer field
\item S.Ds: vector; standard deviation of each integer field
\item Skews: vector; pearson second coefficient of skewness value for each field
}
\enumerate{
\item Float_Characteristics: list; A list containing the same analysis as above however for the continuous numeric data fields
\item Integer_Data: dataframe; containing the data for the integer fields analysed
\item Double_Data: dataframe; containing the data for the continuous numeric fields analysed
\item GT_20pcnt_Flag: list; A list containing a vector for each field analysed where the vector contains the names of each value that makes up over 20\% of the values seen, a potential indicator of either an alternate NA value or what is actually categorical data
}
}
\description{
This function receives a dataframe containing survey data with one column
for each question and will output summary and characteristic values for the
numeric data. Analysis is split in to integer and continuous types. Columns
can either be automatically detected or forced by the user for analysis.
}
\examples{
Numeric_Uni_EDA(dataset = example_dataset,
                    numeric.force = c('col5','col7'),
                    ignore.columns = c('col4'),
                    alternate.nas = list(c('col1', 'error'))

Numeric_Uni_EDA(example_dataset)

plot(Numeric_Uni_EDA(example_dataset))

print(Numeric_Uni_EDA(example_dataset))

summary(Numeric_Uni_EDA(example_dataset))

}
