% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Uni_EDA.R
\name{plot.aggregate_NLP}
\alias{plot.aggregate_NLP}
\title{Plot function for NLP_Uni_EDA S3 Object}
\usage{
\method{plot}{aggregate_NLP}(x, cor_cut = 0.2)
}
\arguments{
\item{x}{aggregate_NLP s3 object}

\item{cor_cut}{The corrolation score to cut off for connections when plotting - default set to 0.2}
}
\value{
Barplots for sentiment scores, word frequency, bigrams and Trigrams as well as a graph map displaying clusters of corrolated words
}
\description{
s3 method to plot Barplots for sentiment scores, word frequency, bigrams and Trigrams as well as a graph map displaying clusters of corrolated words
}
\examples{
x <- NLP_Uni_EDA(basic_test_data)
plot(x)

}
