% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Uni_EDA.R
\name{plot.Categorical_EDA}
\alias{plot.Categorical_EDA}
\title{Plot function for Categorical_EDA S3 Object}
\usage{
\method{plot}{Categorical_EDA}(x)
}
\arguments{
\item{x}{Categorical_EDA s3 object}
}
\value{
Proportional bar plots of each category with caetgories containing less than 2\% of data grouped as other.
}
\description{
s3 method to plot proportional barplots of categorical data. Categories that make up less than 2\% of the data are grouped together
}
\examples{
x <- Categorical_Uni_EDA(basic_test_data)
plot(x)

}
