% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Describe.R
\name{plot.data_describe}
\alias{plot.data_describe}
\title{Plot method for data_describe s3 class}
\usage{
\method{plot}{data_describe}(x)
}
\arguments{
\item{x}{data_describe s3 object}
}
\description{
Plots pie charts for missingness and data types present in data as well as a missingness plot showing the distribution of missing data throughout the dataset
}
