% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Column_Analysis.R
\name{plot.NLP_Column_Analysis}
\alias{plot.NLP_Column_Analysis}
\title{Plot function for NLP_Column_Analysis S3 Object}
\usage{
\method{plot}{NLP_Column_Analysis}(x, cor_cut = 0.2)
}
\arguments{
\item{x}{NLP_Column_Analysis s3 object}

\item{cor_cut}{The corrolation score to cut off for connections when plotting - default set to 0.2}
}
\value{
Barplots for sentiment scores, word frequency, bigrams and Trigrams as well as a graph map displaying clusters of corrolated words
}
\description{
s3 method to plot Barplots for sentiment scores, word frequency, bigrams and Trigrams as well as a graph map displaying clusters of corrolated words
}
\examples{
x <- NLP_Column_Analysis(basic_test_data$words)
plot(x)

}
