% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Uni_EDA.R
\name{plot.Numeric_EDA}
\alias{plot.Numeric_EDA}
\title{Plot function for Numeric_Uni_EDA S3 Object}
\usage{
\method{plot}{Numeric_EDA}(x)
}
\arguments{
\item{x}{Numeric_EDA s3 object}
}
\value{
Histograms and boxplots for integer data and violin and boxplots for conitnuous numeric type data
}
\description{
s3 method to plot histograms and boxplots for integer data and violin and boxplots for conitnuous numeric type data
}
\examples{
x <- Numeric_Uni_EDA(basic_test_data)
plot(x)

}
