% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Uni_EDA.R
\name{print.aggregate_NLP}
\alias{print.aggregate_NLP}
\title{Print function for NLP_Uni_EDA S3 Object}
\usage{
\method{print}{aggregate_NLP}(x)
}
\arguments{
\item{x}{aggregate_NLP s3 object}
}
\description{
Print function for aggregate_NLP s3 object which displays the fields that have been analysed and the data available for each.
}
