% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Uni_EDA.R
\name{print.Categorical_EDA}
\alias{print.Categorical_EDA}
\title{Print function for Categorical_EDA S3 Object}
\usage{
\method{print}{Categorical_EDA}(x)
}
\arguments{
\item{x}{Categorical_EDA s3 object}
}
\value{
Markdown format summary tables for each of nominal and ordinal data displaying the most and least common categories and their respective counts
}
\description{
s3 method to print summary tables for each of nominal and ordinal data displaying the most and least common categories and their respective counts
}
\examples{
x <- Categorical_Uni_EDA(basic_test_data)
print(x)

}
