% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Describe.R
\name{print.data_describe}
\alias{print.data_describe}
\title{Print method for data_describe s3 class}
\usage{
\method{print}{data_describe}(x)
}
\arguments{
\item{x}{data_describe s3 object}
}
\description{
Prints dimensions, general statistics and missingness groups for data on markdown format
}
