% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type_detect.R
\name{print.data_detected}
\alias{print.data_detected}
\title{Print function for data_detected s3 object}
\usage{
\method{print}{data_detected}(x)
}
\arguments{
\item{x}{data_detected s3 object}
}
\description{
Prints the top 10 rows of the output data following running the data_type_detect() function
}
