% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Uni_EDA.R
\name{print.Numeric_EDA}
\alias{print.Numeric_EDA}
\title{Print function for Numeric_Uni_EDA S3 Object}
\usage{
\method{print}{Numeric_EDA}(x)
}
\arguments{
\item{x}{Numeric_EDA s3 object}
}
\value{
Markdown format summary tables showing which fields have been analysed as integer or float type data as well as any values flagged as being over 20\% of the values present
}
\description{
s3 method to print summary tables showing which fields have been analysed as integer or float type data as well as any values flagged as being over 20\% of the values present
}
\examples{
x <- Numeric_Uni_EDA(basic_test_data)
print(x)

}
