% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type_detect.R
\name{summary.data_detected}
\alias{summary.data_detected}
\title{Summary function for data_detected s3 object}
\usage{
\method{summary}{data_detected}(x)
}
\arguments{
\item{x}{data_detected s3 object}
}
\description{
Prints markdown friendly table of the input data format detected as well as the output format of data following running the data_type_detect() function
}
