% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Uni_EDA.R
\name{summary.Numeric_EDA}
\alias{summary.Numeric_EDA}
\title{Summary function for Numeric_Uni_EDA S3 Object}
\usage{
\method{summary}{Numeric_EDA}(x)
}
\arguments{
\item{x}{Numeric_EDA s3 object}
}
\value{
Markdown format summary tables showing detailed summary tables showing the statistical characteristics of the data
}
\description{
s3 method to show detailed summary tables showing the statistical characteristics of the data as calculated by the function
}
\examples{
x <- Numeric_Uni_EDA(basic_test_data)
summary(x)

}
