% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nominal_Detect.R
\name{Nominal_Detect}
\alias{Nominal_Detect}
\title{Nominal Data Type Detection and Forcing}
\usage{
Nominal_Detect(input_vector, dataset, preserve_nonconform = T, force = F)
}
\arguments{
\item{force}{}
}
\description{
This function receives a dataframe and outputs the dataframe with nominal columns
either forced to nominal or detected and converted to nominal. The user can
force columns to a certain data type via specifying manually or allow a
number of tests on data characteristics to automatically detect the likely
data type. In the common case for survey data where users
can select categories or enter free text and the data is included in one
column, the free text can be split out and preserved in additional columns
that will be added to the dataset.
}
